+++ CHANGELOG.md
@@ -6,6 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unrealeased]
 
+## [0.6.5]
+
+### New
+- Special notification when user join an organisation (with link to contract) or is removed from an organisation.
+
+### Changed
+- pushHistory is done in the notifications daemon, in the PushEventNotification function.
+- dql.Count method return errors instead of exiting.
+- [notification] UserCtx now integrate the User.name field, and in email notifications.
+- [notification] notifications for contract are now hanledoutside the resolvers.
+
+### Fixed
+- [auth] Dgraph token crashed if user was first_link of a circle.
+- [auth] HasCoordos ensure that look for coordo role with first_link.
+- [auth] protect membership role (Guest, Owner etc) and allow users to be remove from an organisation.
+- [auth] changing userCanJoin settings correctly set the children visibility.
+- [notification] fixed how we retrive first_link for ReasonIsFirstLink.
+- [notification] fix notification reasons order.
+
 
 ## [0.6.4]